requirements.txt
discord.py
google-genai
python-dotenv
Pillow
aiohttp